fixing block refresh cache method
git-svn-id: http://svn.concrete5.org/svn/concrete5@421 b0551a0c-1e16-4222-a7d5-975db1aca215

+++ trunk/web/concrete/models/block.php
@@ -800,7 +800,9 @@ defined('C5_EXECUTE') or die(_("Access Denied."));
 		public function refreshCache() {
 			$c = $this->getBlockCollectionObject();
 			$a = $this->getBlockAreaObject();
-			Cache::delete('block', $this->getBlockID() . ':' . $c->getCollectionID() . ':' . $a->getAreaHandle());
+			if (is_object($c) && is_object($a)) { 
+				Cache::delete('block', $this->getBlockID() . ':' . $c->getCollectionID() . ':' . $a->getAreaHandle());
+			}
 		}
 		
 		function updateBlockInformation($data) {
